requirements.txt
appdirs==1.4.4
asgiref==3.3.1
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.8.0
django-heroku==0.0.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
filelock==3.0.12
google-auth==1.27.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.0
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
Pillow==8.3.1
pipenv==2021.5.29
protobuf==3.15.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.6
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
